+++ couchbasekit/__init__.py
@@ -3,8 +3,7 @@
 couchbasekit
 ~~~~~~~~~~~~
 
-couchbasekit is a wrapper around couchbase driver for document validation
-and more.
+A wrapper around CouchBase Python driver for document validation and more.
 
 :website: http://github.com/kirpit/couchbasekit
 :copyright: Copyright 2012, Roy Enjoy <kirpit *at* gmail.com>, see AUTHORS.txt.
@@ -12,10 +11,12 @@ and more.
 """
 from couchbasekit.connection import Connection
 from couchbasekit.document import Document
+from couchbasekit.viewsync import register_view
 
 __version__ = '0.2.0-dev'
 
 __all__ = (
     Connection,
     Document,
+    register_view,
 )